Remove KeyValueStreamDataChannel.putBlockRequest and other unused code.
